Moved NavigationDrawer base classes to their own package - Made `NavigationDrawerFragment` abstract so implementing   classes need to specify the menu items themselves
