+++ app/src/main/java/com/poliveria/apps/navigationdrawer/NavigationDrawerCallbacks.java
@@ -1,4 +1,4 @@
-package com.poliveira.apps.materialtests;
+package com.poliveria.apps.navigationdrawer;
 
 /**
  * Created by poliveira on 27/10/2014.
