+++ app/src/main/java/com/poliveria/apps/navigationdrawer/NavigationItem.java
@@ -1,4 +1,4 @@
-package com.poliveira.apps.materialtests;
+package com.poliveria.apps.navigationdrawer;
 
 import android.graphics.drawable.Drawable;
