+++ app/src/main/res/layout/activity_main_topdrawer.xml
@@ -24,7 +24,7 @@
     <!-- android:layout_marginTop="?android:attr/actionBarSize"-->
     <fragment
         android:id="@+id/fragment_drawer"
-        android:name="com.poliveira.apps.materialtests.NavigationDrawerFragment"
+        android:name="com.poliveira.apps.materialtests.MainNavigationDrawerFragment"
         android:layout_width="@dimen/navigation_drawer_width"
         android:layout_height="match_parent"
         android:layout_gravity="start"
